Refactored

+++ src/js/page.js
@@ -1,8 +1,7 @@
 /* eslint-disable no-param-reassign */
 import Masonry from 'masonry-layout';
-import Modal from './modal';
-import Preview from './preview';
-import { animateModals, render } from './utils';
+import Form from './form';
+import { render, subscribeOnNotifications } from './utils';
 
 export default class Page {
   constructor(serverHost, store) {
@@ -12,61 +11,23 @@ export default class Page {
     this.burger = this.page.querySelector('.navbar-burger');
     this.burgerMenu = this.page.querySelector('.navbar-menu');
     this.menuButton = this.page.querySelector('.menu-button');
-    this.add = this.page.querySelector('.add');
-    this.notes = this.page.querySelector('.notes');
     this.emptyList = this.page.querySelector('.notes-empty-list');
     this.notesList = this.page.querySelector('.notes-list');
-    this.addButton = this.page.querySelector('button.notes-go-to-add-button');
-    this.backButton = this.page.querySelector('button.back-button');
     this.audioButton = this.page.querySelector('button.audio-button');
     this.videoButton = this.page.querySelector('button.video-button');
     this.textButton = this.page.querySelector('button.text-button');
-    this.background = this.page.querySelectorAll('section, footer');
+  }
 
+  async init() {
     const masonry = new Masonry(this.notesList, {
       itemSelector: '.notes-list-item',
       columnWidth: '.notes-list-item',
     });
 
-    this.modal = new Modal(this.serverHost, this.page, this.store, masonry);
-    // this.footerLogo = this.page.querySelector('.footer-logo');
-    // this.about = this.page.querySelector('.about');
-
-    // At first, it is needed to subscribe on server notifications
-    // Without this feature, the app freezes when a big file is sent
-    const eventSource = new EventSource(`${this.serverHost}/chest-of-notes/notifications`, { withCredentials: true });
-
-    eventSource.addEventListener('open', () => { console.log('Subscribing on notifications...'); });
-    eventSource.addEventListener('deny', () => {
-      const message = 'Server denied to subscribe on notifications: somebody has already connected';
-      console.log(message);
-      eventSource.close();
-      [this.textButton, this.audioButton, this.videoButton].forEach((button) => {
-        button.disabled = true;
-        button.parentElement.classList.add('hidden');
-      });
-      this.notesList.querySelectorAll('.delete-note').forEach((deleteButton) => {
-        deleteButton.disabled = true;
-        deleteButton.querySelector('svg').style.fill = '#aaaaaa';
-      });
-      setTimeout(() => { alert(message); }, 500);
-    });
-    eventSource.addEventListener('error', () => { console.log('An error occurred with a notifications\' subscription'); });
-    eventSource.addEventListener('uploaderror', (event) => {
-      const message = `An error occurred with a file from the note "${event.data}". Try to record it again`;
-      console.log(message);
-      alert(message);
-    });
-    eventSource.addEventListener('uploadsuccess', (event) => {
-      console.log(`Successfully saved a file from the note "${event.data}"!`);
-      const descriptionToEdit = this.notesList.querySelector(`[data-id="${event.lastEventId}"] .notes-list-item-description`);
-      console.log(descriptionToEdit);
-      const deleteNote = descriptionToEdit.closest('li').querySelector('.delete-note');
-      deleteNote.querySelector('i').style.color = '';
-      deleteNote.disabled = false;
-      descriptionToEdit.textContent = 'Click to open the media!';
-      descriptionToEdit.classList.add('media-content');
-      descriptionToEdit.disabled = false;
+    const unsubscribe = this.store.subscribe(() => {
+      console.log('store:', this.store.getState());
+      masonry.layout();
+      const timeout = setTimeout(() => { masonry.layout(); clearTimeout(timeout); }, 500);
     });
 
     // Listener functions are initiated in a constructor and are given as callbacks
@@ -82,39 +43,10 @@ export default class Page {
         masonry.remove(itemToDelete);
         this.store.setState((previous) => ({ ...previous, items: previous.items.filter((item) => item.id !== dataId) }));
       }
-      // if (!this.notesList.children.length) {
-      //   [this.emptyList, this.notesList].forEach((item) => item.classList.toggle('hidden'));
-      //   this.emptyList.style.visibility = 'visible';
-      // }
     };
 
-    // this.previewListener = (dataType, dataContent, dataId) => {
-    //   const previewWrapper = this.page.querySelector('.preview');
-    //   const preview = new Preview(this.serverHost, previewWrapper, dataId, dataType, dataContent);
-    //   animateModals(previewWrapper, this.background, 'open');
-    //
-    //   const closeListener = (event) => {
-    //     if (event.target.classList.contains('preview')) {
-    //       preview.closeModal(this.background);
-    //       previewWrapper.removeEventListener('click', closeListener);
-    //     }
-    //   };
-    //   previewWrapper.addEventListener('click', closeListener);
-    //   // A way to close without a mouse
-    //   previewWrapper.addEventListener('keyup', (event) => {
-    //     if (event.key === 'Escape') {
-    //       previewWrapper.dispatchEvent(new Event('click'));
-    //     }
-    //   }, { once: true });
-    // };
-    const unsubscribe = store.subscribe(() => {
-      console.log('store:', store.getState());
-      masonry.layout();
-      const timeout = setTimeout(() => { masonry.layout(); clearTimeout(timeout); }, 500);
-    });
-
     this.previewListener = (id, button, type, pipeBlob) => {
-      const { opened } = store.getState();
+      const { opened } = this.store.getState();
       const close = (idToClose) => {
         const previousMediaContainer = this.notesList.querySelector(`[data-id="${idToClose}"]`);
         const previousDescription = previousMediaContainer.querySelector('.notes-list-item-description');
@@ -140,30 +72,37 @@ export default class Page {
         media.style.width = '100%';
         media.style.borderRadius = '4px';
         button.insertAdjacentElement('afterend', media);
-        store.setState((previous) => ({ ...previous, opened: id }));
+        this.store.setState((previous) => ({ ...previous, opened: id }));
       } else {
         close(opened);
-        store.setState((previous) => ({ ...previous, opened: null }));
+        this.store.setState((previous) => ({ ...previous, opened: null }));
       }
     };
+
+    const contentButtons = [this.textButton, this.audioButton, this.videoButton];
+    const isSubscribed = await subscribeOnNotifications(this.serverHost, this.notesList);
 
     // At first, fetched notes must be rendered
     const fetchedData = this.store.getState().items;
     // const fetchedData = [];
+    let deleteButtonsAndIcons;
     if (fetchedData.length) {
       this.notesList.classList.remove('is-hidden');
-      fetchedData.forEach((note) => {
-        const { deleteNote: deleteButton, notesListItemDescription: previewButton } = render(
+      deleteButtonsAndIcons = fetchedData.map((note) => {
+        const { deleteNote: deleteButton, notesListItemDescription: previewButton, icon } = render(
           'note',
           this.notesList,
           note,
           null,
           masonry,
         );
-        deleteButton.addEventListener('click', (event) => this.deleteListener(event, note.id), { once: true });
+        if (isSubscribed === 'allow') {
+          deleteButton.addEventListener('click', (event) => this.deleteListener(event, note.id), { once: true });
+        }
         if (previewButton instanceof HTMLButtonElement) {
           previewButton.addEventListener('click', () => this.previewListener(note.id, previewButton, note.type, null));
         }
+        return { button: deleteButton, icon };
       });
       masonry.layout();
       const timeout = setTimeout(() => { masonry.layout(); clearTimeout(timeout); }, 500);
@@ -172,44 +111,22 @@ export default class Page {
       // this.notesList.hidden = false;
     }
 
-    // // Then, add functionality to spoilers
-    // this.page.addEventListener('click', (event) => {
-    //   if (event.target.classList.contains('checkbox')) {
-    //     const spoiler = event.target;
-    //     const description = spoiler.closest('li').querySelector('.notes-list-item-description');
-    //     if (!event.isTrusted && !event.fromKeyboard) {
-    //       spoiler.checked = false;
-    //     }
-    //     if (spoiler.checked) {
-    //       description.style.maxHeight = `${description.scrollHeight}px`;
-    //       Array.from(this.page.querySelectorAll('.checkbox'))
-    //         .filter((item) => item.checked && item !== spoiler)
-    //         .forEach((item) => { item.dispatchEvent(new Event('click', { bubbles: true })); });
-    //     } else {
-    //       description.style.maxHeight = '';
-    //     }
-    //     const timeout = setTimeout(() => { masonry.layout(); clearTimeout(timeout); }, 600);
-    //   }
-    // });
-
-    // // Expanded notes should adapt if layout changes
-    // window.addEventListener('resize', () => {
-    //   const timeout = setTimeout(() => {
-    //     Array.from(this.page.querySelectorAll('.checkbox'))
-    //       .filter((spoiler) => spoiler.checked)
-    //       .forEach((spoiler) => {
-    //         const description = spoiler.closest('li').querySelector('.notes-list-item-description');
-    //         description.style.maxHeight = `${description.scrollHeight}px`;
-    //       });
-    //     masonry.layout();
-    //     clearTimeout(timeout);
-    //   }, 1000);
-    // });
-
-    // this.notes.scrollIntoView();
+    if (isSubscribed !== 'allow') {
+      contentButtons.forEach((button) => {
+        button.disabled = true;
+        button.parentElement.classList.add('hidden');
+      });
+      deleteButtonsAndIcons.forEach((buttonAndIcon) => {
+        buttonAndIcon.button.disabled = true;
+        buttonAndIcon.icon.style.color = '#aaaaaa';
+      });
+    } else {
+      this.addEventListeners(masonry, contentButtons);
+    }
+    deleteButtonsAndIcons = null;
   }
 
-  addEventListeners() {
+  addEventListeners(masonry, contentButtons) {
     /**
      * A listener for showing adding options (desktop)
      */
@@ -227,36 +144,9 @@ export default class Page {
       this.burgerMenu.classList.toggle('is-active');
     });
 
-    //   /**
-    //    * A listener for showing an 'about' modal
-    //    */
-    //   this.footerLogo.addEventListener('click', () => this.about.classList.add('active'));
-    //
-    //   /**
-    //    * A listener for hiding an 'about' modal after the animation end
-    //    */
-    //   this.about.addEventListener('animationend', () => this.about.classList.remove('active'));
-    //
-    //   /**
-    //    * A listener for an 'add' button
-    //    */
-    //   this.addButton.addEventListener('click', (event) => {
-    //     event.preventDefault();
-    //     this.add.scrollIntoView({ behavior: 'smooth' });
-    //   });
-    //
-    //   /**
-    //    * A listener for a 'back' button
-    //    */
-    //   this.backButton.addEventListener('click', (event) => {
-    //     event.preventDefault();
-    //     this.notes.scrollIntoView({ behavior: 'smooth' });
-    //   });
-    //
     /**
      * A listener to open the content modal
      */
-    const contentButtons = [this.audioButton, this.videoButton, this.textButton];
     contentButtons.forEach((button) => {
       // Make a listener for each button
       const listener = (event) => {
@@ -266,15 +156,8 @@ export default class Page {
             btn.dispatchEvent(new Event('click'));
           }
         });
-        this.modal.openForm(this.serverHost, button, contentButtons, this.deleteListener, this.previewListener);
 
-        // // Resolve a modal view according to a clicked button
-        // // eslint-disable-next-line max-len
-        // const { modalAdd, type } = this.modal.openModal(this.serverHost, button, contentButtons, this.deleteListener, this.previewListener);
-        // animateModals(modalAdd, this.background, 'open');
-        // if (type !== 'text') {
-        //   this.modal.addModalButtonListeners();
-        // }
+        this.form = new Form(this.serverHost, this.notesList, this.store, masonry, button.name, this.deleteListener, this.previewListener);
       };
       button.addEventListener('click', listener);
     });
